sharedfunctions
