+++ scripts/config/fpunitdef.txt
@@ -3,9 +3,9 @@ Oil = bbl
 Clay = tons
 Quartz Crystal = tickets/pounds
 Salt = tons
-Geothermal - Electrical Generation = Other
-Geothermal - Electrical Generation = Kilowatt Hours
 Geothermal - Electrical Generation = Thousands of Pounds
+Geothermal - Electrical Generation = Kilowatt Hours
+Geothermal - Electrical Generation = Other
 Geothermal - Direct Utilization = Hundreds of Gallons
 Geothermal - Direct Utilization = Millions of BTUs
 Anhydrous Sodium Sulfate = tons
